refactoring: 패키지 구조명 변경
