+++ src/main/java/com/api/onboardingkit/OnboardingKitBackendApplication.java
@@ -1,4 +1,4 @@
-package com.example.onboardingkitbackend;
+package com.api.onboardingkit;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
