+++ src/main/java/com/api/onboardingkit/article/repository/ArticleRepository.java
@@ -1,6 +1,6 @@
-package com.example.onboardingkitbackend.article.repository;
+package com.api.onboardingkit.article.repository;
 
-import com.example.onboardingkitbackend.article.model.Article;
+import com.api.onboardingkit.article.model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
